remediate some review comments; move to better package
